// frontend/src/lib/generated/api/models/DTOPeticionRegistroProfesor.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Academia API
 * API para gestión de academia con roles de administrador, profesor y alumno
 *
 * The version of the OpenAPI document: 1.0
 * Contact: contact@academia.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface DTOPeticionRegistroProfesor
 */
export interface DTOPeticionRegistroProfesor {
    /**
     * 
     * @type {string}
     * @memberof DTOPeticionRegistroProfesor
     */
    username: string;
    /**
     * 
     * @type {string}
     * @memberof DTOPeticionRegistroProfesor
     */
    password: string;
    /**
     * 
     * @type {string}
     * @memberof DTOPeticionRegistroProfesor
     */
    firstName: string;
    /**
     * 
     * @type {string}
     * @memberof DTOPeticionRegistroProfesor
     */
    lastName: string;
    /**
     * 
     * @type {string}
     * @memberof DTOPeticionRegistroProfesor
     */
    dni: string;
    /**
     * 
     * @type {string}
     * @memberof DTOPeticionRegistroProfesor
     */
    email: string;
    /**
     * 
     * @type {string}
     * @memberof DTOPeticionRegistroProfesor
     */
    phoneNumber?: string;
    /**
     * 
     * @type {Array<number>}
     * @memberof DTOPeticionRegistroProfesor
     */
    classIds?: Array<number>;
}

/**
 * Check if a given object implements the DTOPeticionRegistroProfesor interface.
 */
export function instanceOfDTOPeticionRegistroProfesor(value: object): value is DTOPeticionRegistroProfesor {
    if (!('username' in value) || value['username'] === undefined) return false;
    if (!('password' in value) || value['password'] === undefined) return false;
    if (!('firstName' in value) || value['firstName'] === undefined) return false;
    if (!('lastName' in value) || value['lastName'] === undefined) return false;
    if (!('dni' in value) || value['dni'] === undefined) return false;
    if (!('email' in value) || value['email'] === undefined) return false;
    return true;
}

export function DTOPeticionRegistroProfesorFromJSON(json: any): DTOPeticionRegistroProfesor {
    return DTOPeticionRegistroProfesorFromJSONTyped(json, false);
}

export function DTOPeticionRegistroProfesorFromJSONTyped(json: any, ignoreDiscriminator: boolean): DTOPeticionRegistroProfesor {
    if (json == null) {
        return json;
    }
    return {
        
        'username': json['username'],
        'password': json['password'],
        'firstName': json['firstName'],
        'lastName': json['lastName'],
        'dni': json['dni'],
        'email': json['email'],
        'phoneNumber': json['phoneNumber'] == null ? undefined : json['phoneNumber'],
        'classIds': json['classIds'] == null ? undefined : json['classIds'],
    };
}

export function DTOPeticionRegistroProfesorToJSON(json: any): DTOPeticionRegistroProfesor {
    return DTOPeticionRegistroProfesorToJSONTyped(json, false);
}

export function DTOPeticionRegistroProfesorToJSONTyped(value?: DTOPeticionRegistroProfesor | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'username': value['username'],
        'password': value['password'],
        'firstName': value['firstName'],
        'lastName': value['lastName'],
        'dni': value['dni'],
        'email': value['email'],
        'phoneNumber': value['phoneNumber'],
        'classIds': value['classIds'],
    };
}