// frontend/src/lib/generated/api/models/DTOActualizacionProfesor.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Academia API
 * API para gestión de academia con roles de administrador, profesor y alumno
 *
 * The version of the OpenAPI document: 1.0
 * Contact: contact@academia.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Professor update data
 * @export
 * @interface DTOActualizacionProfesor
 */
export interface DTOActualizacionProfesor {
    /**
     * 
     * @type {string}
     * @memberof DTOActualizacionProfesor
     */
    firstName?: string;
    /**
     * 
     * @type {string}
     * @memberof DTOActualizacionProfesor
     */
    lastName?: string;
    /**
     * 
     * @type {string}
     * @memberof DTOActualizacionProfesor
     */
    dni?: string;
    /**
     * 
     * @type {string}
     * @memberof DTOActualizacionProfesor
     */
    email?: string;
    /**
     * 
     * @type {string}
     * @memberof DTOActualizacionProfesor
     */
    phoneNumber?: string;
    /**
     * 
     * @type {Set<string>}
     * @memberof DTOActualizacionProfesor
     */
    classIds?: Set<string>;
    /**
     * 
     * @type {boolean}
     * @memberof DTOActualizacionProfesor
     */
    enabled?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof DTOActualizacionProfesor
     */
    empty?: boolean;
}

/**
 * Check if a given object implements the DTOActualizacionProfesor interface.
 */
export function instanceOfDTOActualizacionProfesor(value: object): value is DTOActualizacionProfesor {
    return true;
}

export function DTOActualizacionProfesorFromJSON(json: any): DTOActualizacionProfesor {
    return DTOActualizacionProfesorFromJSONTyped(json, false);
}

export function DTOActualizacionProfesorFromJSONTyped(json: any, ignoreDiscriminator: boolean): DTOActualizacionProfesor {
    if (json == null) {
        return json;
    }
    return {
        
        'firstName': json['firstName'] == null ? undefined : json['firstName'],
        'lastName': json['lastName'] == null ? undefined : json['lastName'],
        'dni': json['dni'] == null ? undefined : json['dni'],
        'email': json['email'] == null ? undefined : json['email'],
        'phoneNumber': json['phoneNumber'] == null ? undefined : json['phoneNumber'],
        'classIds': json['classIds'] == null ? undefined : new Set(json['classIds']),
        'enabled': json['enabled'] == null ? undefined : json['enabled'],
        'empty': json['empty'] == null ? undefined : json['empty'],
    };
}

export function DTOActualizacionProfesorToJSON(json: any): DTOActualizacionProfesor {
    return DTOActualizacionProfesorToJSONTyped(json, false);
}

export function DTOActualizacionProfesorToJSONTyped(value?: DTOActualizacionProfesor | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'firstName': value['firstName'],
        'lastName': value['lastName'],
        'dni': value['dni'],
        'email': value['email'],
        'phoneNumber': value['phoneNumber'],
        'classIds': value['classIds'] == null ? undefined : Array.from(value['classIds'] as Set<any>),
        'enabled': value['enabled'],
        'empty': value['empty'],
    };
}