// frontend/src/lib/generated/api/models/DTOProfesorPublico.ts
export interface DTOProfesorPublico {
    /**
     * Unique professor ID
     * @type {number}
     * @memberof DTOProfesorPublico
     */
    id: number;
    /**
     * Professor's first name
     * @type {string}
     * @memberof DTOProfesorPublico
     */
    firstName: string;
    /**
     * Professor's last name
     * @type {string}
     * @memberof DTOProfesorPublico
     */
    lastName: string;
    /**
     * Professor's email
     * @type {string}
     * @memberof DTOProfesorPublico
     */
    email: string;
    /**
     * Phone number
     * @type {string}
     * @memberof DTOProfesorPublico
     */
    phoneNumber?: string;
    /**
     * Indicates if the account is enabled
     * @type {boolean}
     * @memberof DTOProfesorPublico
     */
    enabled: boolean;
    /**
     * Number of assigned classes
     * @type {number}
     * @memberof DTOProfesorPublico
     */
    classCount: number;
    /**
     * Profile creation date
     * @type {Date}
     * @memberof DTOProfesorPublico
     */
    createdAt?: Date;
    /**
     * 
     * @type {string}
     * @memberof DTOProfesorPublico
     */
    fullName?: string;
    /**
     * 
     * @type {string}
     * @memberof DTOProfesorPublico
     */
    enabledStatus?: string;
}